make Phoenix truly optional
Prior to this commit when `metrix` was brought into a project as a
dependency that did _not_ have Phoenix as a dependency, `metrix` would
fail to compile because it was unable to `import Phoenix.Controller`.
This commit removes that import and fully qualifies the function calls
to `Phoenix.Controller.action_name` and
`Phoenix.Controller.controller_module` so that dependent projects can
still rely on `metrix`, but don't need to bring `phoenix` in as well.